% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circropper.R
\name{circropper}
\alias{circropper}
\title{Crop the image to a circle}
\usage{
circropper(input_path, output_path, margin)
}
\arguments{
\item{input_path}{character the image file path}

\item{margin}{double}

\item{outut_path}{character the saved image path}
}
\value{
an image file saved in the designated folder
}
\description{
Create a new image cropped by a circle mask and leave a margin
}
\examples{
circropper("path_to_input_img.png", "output_img.png", 20)
}
