% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducolor.R
\name{redu_color}
\alias{redu_color}
\title{Reduce image colors}
\usage{
redu_color(input_path, ncolor, output_path = NULL)
}
\arguments{
\item{input_path}{character the file path of the image}

\item{ncolor}{integer number of colors to be reduced to}

\item{output_path}{character if not None, the modified image will be saved
in the provided folder path}
}
\value{
modified image array
}
\description{
Reduce image colors
}
\examples{
redu_color("images/milad.jpg", "images", 4)
}
