% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{A function to reduce the dimension of a give image by removing vertical and horizontal seams}
\usage{
reduce_dimensions(input_file, output_file, width, height)
}
\arguments{
\item{input_file}{character}

\item{output_file}{character}
}
\value{
an array of image
}
\description{
A new image with the modified width and height will be created
}
\examples{
reduce_dimensions(input_file,output_file,width,height)
}
