% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducolor.R
\name{reducolor}
\alias{reducolor}
\title{Reduce image color to either 2 or 8 colors for cartoonized effect}
\usage{
reducolor(style, input_path, output_path = NULL)
}
\arguments{
\item{style}{integer either 0 or 1
0 for white and black colors
1 for 8 colors}

\item{input_path}{character the image file path}

\item{output_path}{character if not Null, the modified image will be saved
in the provided folder path and name}
}
\value{
modified image array
}
\description{
Reduce image color to either 2 or 8 colors for cartoonized effect
}
\examples{
redu_color(0, "tests/milad.jpg", "tests/milad_new.jpg")
}
