% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circropper.R
\name{circropper}
\alias{circropper}
\title{Crop the image to a circle}
\usage{
circropper(x)
}
\arguments{
\item{input_path}{character}

\item{outut_path}{character}
}
\value{
an array of image
}
\description{
Create a new image array cropped by a circle mask
}
\examples{
circropper("path_to_input_img.png", "output_img.png")
}
