% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circropper.R
\name{circropper}
\alias{circropper}
\title{Crop the image to a circle}
\usage{
circropper(input_path, margin)
}
\arguments{
\item{input_path}{character the image file path}

\item{margin}{double the distance between circle boundary and the original image boundary}
}
\value{
a cropped image
}
\description{
Create a new image cropped by a circle mask and leave a margin
}
