% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducolor.R
\name{reducolor}
\alias{reducolor}
\title{Reduce image color to either 2 or 8 colors for cartoonized effect}
\usage{
reducolor(input_path, style, output_path = NULL)
}
\arguments{
\item{input_path}{character the image file path}

\item{style}{string vector
selected two colors from c("white", "black", "red", "green", "blue", "yellow", "pink", "aqua" )
or "eight" for eight colors}

\item{output_path}{character if not Null, the modified image will be saved
in the provided folder path and name}
}
\value{
modified image array
}
\description{
Reduce image color to either 2 or 8 colors for cartoonized effect
}
\examples{
input_path <- system.file("tmp_image", "mandrill.jpg", package = "rimager")
new <- reducolor(input_path, c("black", "white"), "new.jpg")
OpenImageR::imageShow(new)
}
