% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redusize.R
\name{redusize}
\alias{redusize}
\title{A function to reduce the dimension of a give image by removing vertical and horizontal seams}
\usage{
redusize(input_file, output_file, width, height)
}
\arguments{
\item{input_file}{character, the file path for the input file}

\item{output_file}{character, the file path for the output file}

\item{width}{integer, the width of the reduced image}

\item{height}{integer, the height of the reduce image}
}
\value{
an array of image
}
\description{
A new image with the modified width and height will be created
}
\examples{
input_path <- system.file("tmp_image", "mandrill.jpg", package = "rimager")
new <- redusize(input_path, "reduced.jpg", 297, 200)
}
